+++ src/mbgl/renderer/tile_pyramid.hpp
@@ -15,6 +15,7 @@
 #include <unordered_map>
 #include <vector>
 #include <map>
+#include <set>
 
 namespace mbgl {